+++ NEWS.md
@@ -5,8 +5,7 @@
 * `createStudy()` integrate "st-storage".
 * `createArea()` integrate "st-storage".
 * `removeArea()`  integrate "st-storage" and "mingen" data.
-* `writeHydroValues()` integrate "mingen" data and dependency with "mod.txt" data.
-* `writeInputTS()` integrate dependency with "mingen" data.
+* `writeInputTS()` integrate "mingen" data and dependency between "mod.txt" and "mingen.txt" data.
 * `createCluster()` / `editCluster()` have new parameter `list_polluants ` for list of pollutants.